Fix floating ip port mapping for multiple ports
When floating ip's are mapped to multiple ports (using config api),
this is not represented to neutron correctly.

When using octavia (or when creating a HA scenario), a port is created
for the VIP, and this VIP is used as allowed-address-pair on the other
ports.

This fix will select the VIP port if possible, so octavia still knows
which floating ip is assigned to the load balancer.

The selection process is using the instance-ip of a port to see if the
fixed_ip_address of the floating ip matches with the port.